+++ src/server/remote/remote-server.ts
@@ -10,6 +10,7 @@ import { TemporarySet } from '../../misc/temporary-set'
 import { runTask } from 'cc-instanceinator/src/inst-util'
 import { sendRemoteServerPacket } from './remote-server-sender'
 import { PhysicsUpdatePacketEncoderDecoder } from '../../net/binary/physics-update-packet-encoder-decoder.generated'
+import { applyModCompatibilityList, ModCompatibilityList } from '../mod-compatibility-list'
 
 import './ignore-pause-screen'
 import './entity-physics-forcer'
@@ -25,6 +26,7 @@ export interface RemoteServerConnectionSettings {
 
 export interface RemoteServerSettings extends ServerSettings {
     connection: RemoteServerConnectionSettings
+    modCompatibility?: ModCompatibilityList
 }
 
 export interface ClientLeaveData {
@@ -50,6 +52,8 @@ export class RemoteServer extends Server<RemoteServerSettings> {
         if (!REMOTE) return
 
         await super.start()
+
+        if (this.settings.modCompatibility) applyModCompatibilityList(this.inst, this.settings.modCompatibility)
 
         const connS = this.settings.connection
         if (connS.type == 'socket') {