+++ src/types.ts
@@ -3,6 +3,7 @@ import type {} from 'crossnode/crossnode'
 import type {} from 'ccmodmanager/types/plugin'
 import type {} from 'ccmodmanager/types/gui/menu'
 import type {} from 'nax-ccuilib/src/ui/pause-screen/pause-screen-api'
+import type {} from 'nax-ccuilib/src/ui/quick-menu/quick-menu-extension'
 import 'ts-binarifier/src/type-aliases'
 
 declare global {